Fix and improve codebase based on reviewer feedback, reflect final logic
   Sensitivity analysis:
   - Fix DGSM normalization: use SALib's built-in dgsm output (nu/pi^2*sigma^2)
     directly for "dgsm" mode; only apply sigma_x/sigma_y for "vi" mode
   - Fix bug where normalized values were overwritten with raw values
     (line 106 re-read sp.analysis instead of using computed s1)
   - Add aggregation support (--aggregation none/quantity/region/year) with
     z-score pre-standardization across non-target dimensions
   - Add sigma_normalization_factor helper with epsilon clamping for flat outputs
   - Add _finite_r2 helper to handle NaN/Inf in sensitivity R2 tables
   - Update deprecated polars groupby -> group_by

   Other bug fixes:
   - Fix r2_score argument order: r2_score(y_true, y_pred) not (y_pred, y_true)
   - Fix NotImplentedError typo -> NotImplementedError
   - Fix scenario_id Int16 -> Int32 to avoid overflow

+++ ml_gcam/data/targets.py
@@ -55,7 +55,7 @@ SOURCES_WITH_DATA = Source.sampled()
 ALL_SPLITS = Split.all()
 ID_COLS = [
     pl.col("experiment").cast(pl.String),
-    pl.col("scenario_id").cast(pl.Int16),
+    pl.col("scenario_id").cast(pl.Int32),
     pl.col("region").cast(pl.String),
     pl.col("year").cast(pl.Int16),
     pl.col("split").cast(pl.String),
